Add an SCF skin to style SCF components to match your site's look and feel
Add a style sheet the overrides the default SCF styles for those
aspects of the components that you want to change. The SCF selectors
are well namespaced, descriptive and modular to make necessary changes
easily.
Add the skin to your site design so that it gets applied on your pages.