Add an Ideation component that extends the AEM Communities Forum - HBS component
Create ideation forum, ideation topic (idea), ideation post (replies)
components and set them up so that their sling:resourceSuperType
property points to the corresponding component in the OOTB AEM
Communities forum components.
Copy over the templates from the forum components, this will give us a
base to build the new ideation component UI off of. You only need to
copy over a template if you think you might have to make some changes
to it, if not, Sling will automatically fall back to the super
component and pick up an appropriate template.
Make markup changes to the hbs templates so that the forum is
customized enough to start to look like an ideation forum.