Add a SocialComponent to customize the data used to render an Idea
Add a SocialComponent that extends the AbstractPost class.
Override the getTags method in order to filter out the "status"
specific tags.
Implement a SocialComponentFactory that registers the new
SocialComponent with SCF so that SCF knows to use the new
IdeaSocialComponent every time a resource of type
"acme/components/ideation/topic" is being requested for.